snippets/viewsOld6 - Authentication & Permissions.py
from django.contrib.auth.models import User
from rest_framework import permissions #Required Permissions to views
from snippets.models import Snippet
from snippets.serializers import SnippetSerializer, UserSerializer
from snippets.permissions import IsOwnerOrReadOnly #Customização de permissão
from rest_framework import generics


class UserList(generics.ListAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer

class UserDetail(generics.RetrieveAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer


class SnippetList(generics.ListCreateAPIView):
    permission_classes = (permissions.IsAuthenticatedOrReadOnly, #Required Permissions to views
                          IsOwnerOrReadOnly) #Custom Permission to view
    queryset = Snippet.objects.all()
    serializer_class = SnippetSerializer

    def perform_create(self, serializer):
        serializer.save(owner=self.request.user)


class SnippetDetail(generics.RetrieveUpdateDestroyAPIView):
    permission_classes = (permissions.IsAuthenticatedOrReadOnly, #Required Permissions to views
                          )
    queryset = Snippet.objects.all()
    serializer_class = SnippetSerializer
