requirements.txt
certifi==2018.1.18
Django==2.0.3
djangorestframework==3.7.7
gunicorn==19.7.1
Pygments==2.2.0
pytz==2018.3
wincertstore==0.2
